serverless/predict-app/app/requirements.txt
tensorflow==2.6.0
pillow==8.2.0
tifffile
numpy
pandas
opencv-python
shapely
keras==2.6.0